+++ src/dashboard-refactor/header/SidebarHeader/types.tsx
@@ -1,14 +1,3 @@
-export interface SidebarToggleProps {
-    sidebarLockedState: SidebarLockedState
-    sidebarToggleHoverState: SidebarToggleHoverState
-}
-
-export interface SidebarToggleHoverState {
-    onHoverEnter(): void
-    onHoverLeave(): void
-    isHovered: boolean
-}
-
 export interface SidebarLockedState {
     toggleSidebarLockedState(): void
     isSidebarLocked: boolean
