+++ src/dashboard-refactor/types.tsx
@@ -1,0 +1,5 @@
+export interface HoverState {
+    onHoverEnter(): void
+    onHoverLeave(): void
+    isHovered: boolean
+}
